neural_dynamics/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
grpcio==1.63.0
idna==3.7
Jinja2==3.1.3
kaleido==0.2.1
kiwisolver==1.4.5
lightning==2.2.3
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.21.0
protobuf==5.26.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.3
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
torch==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
Werkzeug==3.0.2
yarl==1.9.4
